Add attr_names and headers to base exporter.
